Modules refactoring and card's pretty preview
