+++ src/components/card/NewCard.tsx
@@ -1,9 +1,6 @@
 import React, { HTMLAttributes } from 'react';
-import { MiniCard } from './MiniCard';
-import { ActiveCardUIModes } from '@/state/slices';
+import { MiniCard } from './visuals';
 import { useAppStore } from '@/state';
-import { useCard } from '@/query/queryHooks';
-import { LoadableComponent } from '../library/LoadableComponent';
 import clsx from 'clsx';
 
 interface NewCardProps extends HTMLAttributes<React.FC> {}
@@ -14,12 +11,13 @@ export const NewCard: React.FC<NewCardProps> = ({ className }) => {
 
   return (
     <MiniCard
-      previewText="+"
       onClick={() => {
         setActiveCardUIMode('edit');
         setDraftActiveCard();
       }}
       className={clsx('bg-1-4 text-7xl font-normal', className)}
-    />
+    >
+      <h2>+</h2>
+    </MiniCard>
   );
 };
