+++ src/components/card/controls/CreateCardControls.tsx
@@ -1,60 +1,46 @@
 import clsx from 'clsx';
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { MultiValue } from 'react-select';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import { Button } from '@/components/library/Button';
-
+import { Button } from '@/components/library';
 import { useAppStore } from '@/state';
 import {
-  useCardCollections,
   getCardQueryOptions,
+  getProfileCardsQueryOptions,
   getCardCollectionsQueryOptions,
 } from '@/query/queryHooks';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { dataExtractionWrapper } from '@/query';
-import { updateCardCardsCardIdPut } from '@/api';
-import { LoadableComponent } from '../library/LoadableComponent';
-import {
-  collectionResponseToOptions,
-  CollectionsSelect,
-  Option,
-} from './CollectionsSelect';
+import { createCardCardsPost } from '@/api';
+import { CollectionsSelect, Option } from './CollectionsSelect';
 
-export const EditCardControls: React.FC = () => {
-  const cardId = useAppStore((state) => state.activeCardId);
+export const CreateCardControls: React.FC = () => {
   const cardData = useAppStore((state) => state.activeCard);
 
-  const { cardCollections, isPending: cardCollectionsPending } =
-    useCardCollections(cardId);
   const [selectedOptions, setSelectedOptions] = useState<
     MultiValue<Option<number>>
   >([]);
 
-  useEffect(() => {
-    setSelectedOptions(collectionResponseToOptions(cardCollections));
-  }, [cardCollections]);
-
   const client = useQueryClient();
-  const { mutate: updateCard, error: updateError } = useMutation({
+  const { mutate: createCard, error: createError } = useMutation({
     mutationFn: (data: {
-      id: number;
-      new_card: { frontSide: string; backSide: string };
+      card: { frontSide: string; backSide: string };
       collections: number[];
     }) =>
       dataExtractionWrapper(
-        updateCardCardsCardIdPut({
-          path: {
-            card_id: data.id,
-          },
+        createCardCardsPost({
           body: {
             ...data,
           },
         })
       ),
     onSuccess: (responseData, requestData) => {
-      client.invalidateQueries({ queryKey: ['collection'] }); // TODO: подумать, как оптимизировать этот запрос
+      client.invalidateQueries({ queryKey: ['collection'] });
       client.invalidateQueries({
         queryKey: getCardCollectionsQueryOptions(responseData.id).queryKey,
+      });
+      client.invalidateQueries({
+        queryKey: getProfileCardsQueryOptions().queryKey,
       });
       client.setQueryData(
         getCardQueryOptions(responseData.id).queryKey,
@@ -77,16 +63,14 @@ export const EditCardControls: React.FC = () => {
           <span className="md:text-right w-full md:w-1/6 px-1">
             Paired with:
           </span>
-          <LoadableComponent isPending={cardCollectionsPending}>
-            <CollectionsSelect
-              selectedOptions={selectedOptions}
-              setSelectedOptions={setSelectedOptions}
-            />
-          </LoadableComponent>
+          <CollectionsSelect
+            selectedOptions={selectedOptions}
+            setSelectedOptions={setSelectedOptions}
+          />
         </div>
-        {updateError && (
+        {createError && (
           <div className={clsx('center mb-2', 'text-red-200 font-bold')}>
-            {updateError.message}
+            {createError.message}
           </div>
         )}
       </div>
@@ -95,14 +79,13 @@ export const EditCardControls: React.FC = () => {
           className="text-xl m-3"
           variant="bordered"
           onClick={() => {
-            updateCard({
-              id: cardId,
-              new_card: { ...cardData },
+            createCard({
+              card: { ...cardData },
               collections: selectedOptions?.map((option) => option.value),
             });
           }}
         >
-          Save
+          Create card
         </Button>
       </div>
     </>
