+++ src/components/card/visuals/index.ts
@@ -1,0 +1,2 @@
+export * from './MiniCard';
+export * from './FlippingCard';
