+++ src/components/editor/index.ts
@@ -1,0 +1,2 @@
+export * from './EditorComponent';
+export * from './MarkdownRenderComponent';
